Horde: Write the name of the channel to use for agent updates into the agent config file, and report it to the server through the agent properties.
Each configuration now runs a command in the agent itself to update its tool id prior to upload. This allows removing opinionated logic from the server, and allows us to distribute a dedicated win64 agent in the installer.

Does not use this value for updating yet. Will add that in a subsequent change to reduce risk.

[FYI] Carl.Bystrom
#jira UE-205447

[CL 31807005 by ben marsh in ue5-main branch]