Generate bytecode for instantiating top-level module code
#rb russell.johnston
#okforversepublic

Top-level code is allowed to have mostly arbitrary expressions in it. For example, you could have `X:int = 1+2` at the top level. Prior to this patch, we had a janky AST interpreter that pattern matched code at the top level. It only worked for things that were constant to the degree it understood. However, it's cleaner to just generate bytecode for top-level code that uses the normal interperter and binds evaluation results to variables.

[CL 31833870 by saam barati in ue5-main branch]