UE-206314 - Crash preloading InheritableComponentHandler with editor async loading
We're hitting this assert because UInheritableComponentHandler::PreloadAll is calling ConditionalPostLoad. However, it's calling this function because it's attempting execute the fixup for component template names. This can be easily worked around if we simply extract the body of UInheritableComponentHandler::PostLoad and put it in a new function, FixComponentTemplateNames, that both PostLoad and PreloadAll can call.

#jira UE-206314
#rb dan.oconnor, Phillip.Kavan
#rnx

[CL 31884689 by dave jones2 in ue5-main branch]