Moved data fixup from PostLoad to Serialize.
UK2Node_EvaluateProxy2::AllocateDefaultPins was calling postload on its proxy, which was causing issues with editor async loading. Specifically, we don't expect any postload calls to be threadsafe unless the UObject returns true for IsPostLoadThreadSafe. Additionally, we've never postloaded Blueprints and their dependencies during a compilation pass (only preloaded). Calling postload here would be a violation of that implicit rule.

We can easily resolve this by moving the data fixup code in UProxyAsset::PostLoad into Serialize, which is where we normally do that anyway. This alleviates the need for a manual postload call since the proxy is guaranteed to be fully preloaded with data fixup applied.

#rb Keith.Yerex
#rnx

[CL 31886282 by dave jones2 in ue5-main branch]