Cameras: add debug drawing framework
The camera system now maintains a storage for "debug blocks", which are a hierarchy of small structures that can print and display debug information on screen. These structures are meant to be standalone, i.e. they contain the values they need for the debug info. This is because we want to record gameplay and store these debug blocks in trace files, to be later replayed or scrubbed back.

Debug blocks are created and filled during evaluation, each frame, but since they are constantly recreated inside the same storage, there are few, or most often zero memory allocations done on any frame after the first few frames of gameplay.

Debug blocks can be defined "by hand", or can be defined using a set of macros that (hopefully) provide automatic member field serialization (again, for later recording/replay support).

A new debug renderer class provides helper functions for printing and drawing the debug info. In later changelists, it will gain more features, such as colored text, drawing clocks and graphs, and more.

The dampen position camera node is the first to implement debug drawing, as a test.

The camera system component registers itself with the engine's debug draw service to provide an actual hook for this debug drawing.

#jira UE-205980

[CL 31921298 by ludovic chabant in ue5-main branch]