Chaos Cloth: Add Self Collision Thickness Map
Thickness is baked into the Spatial hierarchy when using a map. Should be very little overhead when not using a map.

Tested with new and old style solver (with and without offsets).

Also includes: fix to an API change to GetPointNormals. Fix to a crash in EdgeIntersections detection when you find more than the preallocated number of intersections.

#rb kriss.gossart

[CL 31994472 by alex mcadams in 5.4 branch]