Allow changing the LoopInPin on MidiPlayer node while it's playing.
This actually works perfectly. If you turn looping off, it'll continue playing until the end of the midi file no matter how many times it's looped already. And if you turn looping on, it'll pick up from where it should have been looping.

SetupNewMidiFile does the work of updating the loop data and seek if necessary

[REVIEW] [at]charlie.huguenard [at]buzz.burrowes
#jira UE-208404
#rb charlie.huguenard
#tests PIE, HarmonixTests

[CL 32035041 by jake burga in ue5-main branch]