[NNE] NNERuntimeORTDml implementing the RDG interface
Problem:
For running neural network inference as part of a Render Dependency Graph (RDG), NNE has already a runtime based on HLSL compute shaders with wide platform support but limited model support and performance. There is the need of an additional high performant runtime that implements the NNE RDG interface to enable ML applications for real time rendering.

Solution:
Our plugin NNERuntimeORT already includes a DirectML runtime that implements the GPU interface for neural network inference outside the rendering of a frame. We implement the necessary infrastructure to let this backend also implement the RDG interface for in-frame inference, providing a fast and powerful execution provider on DX12 based desktop systems.

#jira UE-207173
#rb nico.ranieri
[FYI] florent.guinier

#tests Non-unity build, Editor/Game all tests



#virtualized

[CL 32080683 by christian regg in ue5-main branch]