Fixed the two biggest causes for flickering when playing back SparseVolumeTextures: Exceeding the available IO bandwidth and reaching the streaming pool size limit.
IO Bandwidth:
The SVT streamer now features a cvar (r.SparseVolumeTexture.Streaming.BandwidthLimit) which is used to scale down the bandwidth which is used to stream in each streaming instance of a SVT. When the sum of the requested bandwidth of all streaming SVTs exceeds this value, a fractional mip level is computed which is the lowest (highest resolution) mip level which the streamer can continuously stream without exceeding the limit. In order to keep track of the individual streaming instances, there is now a new FStreamingInstance class, which replaces the old FStreamingWindow and is used to keep track for playback direction and to compute the requested and in-budget IO bandwidth. Users calling into the SVT streaming system can pass a uint32 "key" which is used to uniquely identify streaming instances across system ticks. This key makes it possible to tell requests apart that request the same frames but belong to different HV components or SVT viewers and might play at different frame rates.

Streaming Pool Size:
Exposed properties on the SVT asset to allow users to configure the streaming pool size. We now also expose per-SVT settings for prefetching in the same way. When these settings are changed at editor time, the SVT is simply reregistered with the system, which makes the new parameters take effect. Ideally we would automatically grow/resize the streaming pool, but for now this seems like a good intermediate solution.

In order to debug the bandwidth throttling, it turned out to be very helpful to have a way of visualizing what the streamer is doing, so this CL also introduces a debug visualization mode that can be enabled with r.SparseVolumeTexture.Streaming.ShowDebugInfo.

Reviewed in review 31990704

#rb Sebastien.Hillaire
#jira UE-208826
#rnx

#ushell-cherrypick of 31956040 by tim.doerries

[CL 32087494 by tim doerries in ue5-main branch]