Reduce log severity for geometry collection proxies that find missing implicits.
With dataflow we can selectively remove geometries in our pipeline which may be in general included by the MaxSimulatedLevel parameter on a geometry collection component. Previously we would emit an error when attempting to create a proxy if it encountered any empty geometries. This change reduces that to Verbose output so we can still activate these logs if needed to detect unexpected missing geometries but doesn't act as a failure case just because the dataflow removed geometries by design

#rb Aidan.McDu

[CL 32089515 by benn gallagher in 5.4 branch]