Path Tracer: Use stochastic blending to handle opacity on glass materials
This simplifies the logic for solid/thin glass greatly and makes it easier to support transparent thin glass via the new SurfaceCoverage material output (introduced in 32048301).

There is a new shading model id for the path tracer for the solid glass case. This and the existing SHADINGMODELID_THIN_TRANSLUCENT now refer only to the glass lobes. SHADINGMODELID_DEFAULT_LIT can be used for the rest which removes some duplicated logic and makes the common cases much faster.


#rb Sebastien.Hillaire

[CL 32089900 by chris kulla in ue5-main branch]