Prepare MidiOut _after_ registering with the MidiClockIn.
This is because registering with the MidiClock triggers a Reset/Seek. And since the StepSequencer responds immediately to MidiClock Reset/Seeks, as it is a PlayCursor attached to the MidiClock this can result in a race condition of having MidiMessages from the last block.

Although, I'm still a bit unsettled by this code. I still don't know that this guarantees we won't hit this issue again. But 3/3 I haven't hit the assert with these changes.

[REVIEW] [at]charlie.huguenard
#rb buzz.burrowes, charlie.huguenard, matthew.booth
#tests Package Build, Harmonix Test, PIE,, Harmonix Tests

[CL 32106269 by jake burga in ue5-main branch]