Fixed body instance filter calculation when its SourceObject member is unset.
Source objects were added to allow non-actor workflows to use queries and bind their objects to the filter for a body. However it's currently assumed to always be valid. That's not always the case - such as with geoemtry collections and with various custom uses of body instance. Rather than just set the source object from geoemtry collection this change implements the correct fallback logic for the actor as owner if the check for the source object fails which could have been a potential breakage point in the body instance API.

#rb michael.bao, vincent.robert

[CL 32106977 by benn gallagher in ue5-main branch]