Take Recorder: Don't make a unique object name based on the newly added component's name because it might clash when the next newly created component.
For example, if an actor is spawning components, Comp_0, Comp_1, Comp_2, when we notice the first spawned component (Comp_0), we will make a new component based on that name. That new component will be named Comp_1 since it's based off of Comp_0. When we notice the second spawned component (Comp_1), it will already exist in the spawnable object template since we added it from the first component.

#jira UE-208896
#rb matt.hoffman

[CL 32170420 by max chen in ue5-main branch]