Object for homogenising results from IAS components
In a few places in IAS where an action can have multiple outcomes; failure,
pending, disconnect, N bytes received, etc. Some times we use integers which
works well but is not at all clear. There was a prior attempt to tidy this up
(FResult) but it was never seen through to completion. Here we go again, only
with a different name to make life a little less stressful. This submit is just
the object itself.

#rb per.larsson
#rnx

[CL 32174877 by martin ridgers in ue5-main branch]