Make ESC clear filtering of a menu
This fixes an issue where if you first typed into a menu to search and filter it and then pressed ESC, then the query was cleared but the menu was still filtered. After this change, pressing ESC will in the same situation clear the search field and undo any filtering. Also, the search field will revert to its previous visibility, so if it was hidden before the user started typing it will hide after pressing ESC or if it was visible before the user started typing then the search field will just clear but stay visible when ESC is pressed.

#jira UE-200513
#rb aditya.ravichandran

[CL 32183762 by sebastian arleryd in ue5-main branch]