No longer create display strings for unlocalized text
In UE 5.0 we removed the assumption that every FText must have an associated display string. Prior to this change, FText didn't store its TextID locally, and instead resolved it via a reverse LUT against the display string pointer.

While the initial change for 5.0 cleaned up most of the code that was creating redundant display strings, it missed a few places that were still creating display strings for unlocalized text. Now that we can unload the localization data for plugins, it's important that any lingering FText instances don't re-add a redundant display string via FText::Rebuild.

To this end, this change removes the remaining vestiges of the old "every FText must have a display string" assumption and updates FTextLocalizationManager to no longer create display strings for unlocalized text.

This change breaks two pieces of API (which is why it wasn't done initially), however this API was only used within FText itself. The likelihood of this breaking user code is very low.

Breaking changes:
  * FTextLocalizationManager::GetDisplayString now returns a FTextConstDisplayStringPtr rather than a FTextConstDisplayStringRef, and will now return null for an unlocalized text.
  * FTextLocalizationManager::AddDisplayString has been deleted. It has served no purpose since the reverse LUT was removed.

#rb anton.dunchev

[CL 32331936 by jamie dale in ue5-main branch]