[SVG Importer] Imported SVGs textures should be centered
Texture content will now be centered with respect to svg canvas size.

Note: this might still end up having a different result when compared to 3D Geometry generated from the same SVG, for those cases in which SVG shapes are not centered in the original SVG source.
This happens because 3D SVG geometry is centered with respect to its bounding box, so it's not currently taking into account the empty space of the svg viewbox.
E.g. texture generated from SVG defining a 400x400 viewbox containing a 100x100 rectangle placed in 0;0. Something like that will still be centered wrt viewbox, and not the single rectangle it contains. The geometry will be centered wrt the rectangle bounds.


#jira UE-210147

[CL 32370337 by dario mazzanti in ue5-main branch]