[Motion Design] Game Viewport Playable Group Refactor
Summary:

Allow Motion Design playback to occur in a Game Viewport (PIE, Game, etc). Augment the use cases for rundown and playback graphs, including using within nDisplay's workflow.

Refactor Details:

- Playable Group is now more an abstract class. All derived classes have been moved to separate cpp/h.
- The AvaGameInstance implementation has been moved to a derived class: AvaGameInstancePlayableGroup.
- AvaGameViewportPlayableGroup implements the new pre-existing game instance/viewport hook.
- AvaPlayableGroupSubsystem is used to track when a game instance is stopped and unregister it from any playable group it would have been attached to.
- (WIP) PlayableGroupManager: PlayableGroup factory has been updated. Using the presence of GameViewportMediaOutput in the channel to hook to game viewport playable group. This method may be revisited later.

Other:

- GameViewportMediaOutput (WIP):

This is the temporary way we setup a channel to render in a game viewport using dummy GameViewportMediaOutput.
This part is still work in progress.

- Display Cluster Support (WIP):

if the playback server is running as a cluster node, it will append the node's name to the playback server name to be identifiable in the host list in the broadcast editor. This may be a temporary approach until we find a better solution.

- Server name and status in broadcast output tile:

Added more information to indicate which server the output is part of. This had been erroneously removed in a previous refactor.

#rb Juan.Portillo

[CL 32426830 by jean perrier in ue5-main branch]