Automatically classify the most common OOM reasons for UEFN cloud cooking
When we run out of memory during a cloud cook, the most common reason is that users put a large asset (usually mesh or texture) into their project. This CL makes it such that the error for that case gets the full "structured-logging" treatment with key and namespace, so we can easily find it. When we now run out of memory, we'll scan the log for that error (as we do for other errors) and raise it to the user in a friendly way.

Future changes could raise these warnings in the editor already, but with this approach we only warn users when something actually went wrong (and hence avoid false-positives). The limits in the error message are my suggestions for where assets stop being reasonable in the context of memory usage. (Unfortunately, this change specifically does not call out the asset that caused the problem since that would mess with the bucketing in our reporting right now.)

Additionally, this will also declutter our internal reporting and save lots of time previously wasted on investigating failures that are caused by huge assets.
#rb Justin.Marcus

[CL 32474465 by sebastian schoner in 5.4 branch]