Move Razer Chroma plugin to experimental folder.
The Razer Chroma plugin will allow you to import ".chroma" razer animation files (which you can make on the Razer Chroma website and download) and then play them in your game. Making nifty peripheral lighting effects for players responding to your gameplay events.

This is experimental! Expect bugs and changes to the API in the future as it develops.

#jira UE-205451
#rb graham.matuszewski

[CL 32484603 by ben hoffman in ue5-main branch]