Fix issue related to minimizing a notification's parent window
During a short fade in, SNotificationExtendable asks Slate to enter responsive mode. The bug we're fixing here was caused by SNotificationExtendable relying on an SWidget::Tick override to get ticked and eventually release the throttle handle for responsive mode. The bug occurred when the notification's parent window was minimized before the notification is shown, preventint the parent window from ever being painted and thus never ticked, causing the notification to ask for responsive mode but then never ticking out of it, holding onto Slate responsive mode forever. This prevents certain processes in the editor from progressing.

This change includes a related fix of not tying a notification to a window if the window is already minimized or not visible to prevent the notification from showing up as a black square without content because it never gets painted.

#jira UE-207583, UE-206871
#rnx
#rb aditya.ravichandran, Patrick.Boutot, Patrick.Enfedaque

[CL 32489334 by sebastian arleryd in 5.4 branch]