Columns created in TEDS that have intialization data assigned now have that data applied before the OnAdd observer callback is triggered.
Even with the recent change to Mass, columns weren't being assigned their data before OnAdd observers were triggered. This however was a problem on the TEDS side as registring columns happens in a "add" and separate "set" step. These have now be combined togehter into one so the initialization happens before OnAdd is called. For this to work however Mass has gotten an additional function that calls an initialization function if provided, to get the timing correct.

#jira UE-196470
#rb logan.buchy, Mieszko.Zielinski

[CL 32496148 by ronald koppers in 5.4 branch]