Only print SCW memory stats on OOM if lock can be taken immediately
It's useful to get the memory usage of SCWs when an OOM appears. However, when multiple threads encounter an OOM condition at the same time, it is possible that one of them goes to perpertual sleep while holding the lock guarding the SCW data. This means that OOM reporting itself can stall indefinitely trying to take that lock.

#rb dan.elksnitis, Laura.Hermanns

[CL 32498774 by sebastian schoner in 5.4 branch]