Stale data could cause IAS cache to load incorrectly.
When the IAS cache loads, the journal file is compared against the data file to get them back in sync (the former is generally written at a lower frequency than the latter - they can have different views of the state). If the data file cursor was nearing the maximum data file size it would reset the cursor back to zero. This would leave older data at the end of the data file that the journal may still reference. Said false reference would cause cache startup to select the wrong range of the journal to add as items. Now data is written right to the end of the data file so age of bytes in the file always progresses forward.

#rb johan.berg
#rnx

[CL 32538400 by martin ridgers in ue5-main branch]