Intersection tests would occasionally include four bytes too much
One upon a time a four byte identifier was prefixed to a "phrase" of data (a bundle of cache items). This identifier was used to detect discontinuities between the journal and the data as they are not written to attomically. Now we hash the start of a "phrase" so the four byte adjustment for the first item is not applicable.

#rb johan.berg
#rnx

[CL 32538752 by martin ridgers in ue5-main branch]