Fix cases where render resources accumulate despite having a visible realtime viewport.
UpdateSingleViewportClient doesn't guarantee that a viewport will be updated, even if it's realtime (e.g. when FSlateThrottleManager::Get().IsAllowingExpensiveTasks() returns false, which happens if a popup menu is shown). Instead of trying to guess what it did at the call site, this CL makes it return that decision to the caller. (Note that the return value does not mean a viewport was updated, it returns true only if it updated a non-realtime viewport. That should have been an output param as well, but it's too late to change now because it's a public function.)

#jira UE-207385
#rnx
#rb Patrick.Boutot, zach.bethel

[CL 32573845 by mihnea balta in 5.4 branch]